- Renamed package to be underscored instead of camelCase - Enable the simulated robot by default - Removed warnings
