+++ SimpleMockWpiLib/src/com/snobot/simulator/gui/advanced_widgets/SolenoidGraphicDisplay.java
@@ -5,15 +5,12 @@ import java.awt.Graphics;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.border.TitledBorder;
 
-import com.snobot.simulator.gui.Util;
-import com.snobot.simulator.gui.widgetDisplays.BaseWidgetDisplay;
+import com.snobot.simulator.gui.widget_displays.BaseWidgetDisplay;
 
 import edu.wpi.first.wpilibj.Solenoid;
-import edu.wpi.first.wpilibj.SpeedController;
 
 public class SolenoidGraphicDisplay extends BaseWidgetDisplay<Solenoid>
 {
